+++ packages/cap-config-writer/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Minor Changes
 
--   45c01b0: Remove cds version check when adding a cds watch sciprt to root package.json
+-   45c01b0: Remove cds version check when adding a cds watch script to root package.json
 
 ## 0.7.68